refactor(members): deuda estática a CERO — PHPStan sin baseline + PHPCS limpio
Baseline ELIMINADO: PHPStan nivel 5 pasa con 0 errores (ignoreErrors solo
para falsos positivos de harness/stubs WP: get_post/get_userdata nullsafe,
add_submenu_page(null), add_node objeto, require_once /tmp, isset defensivo
en catch, class_exists Gateway en stub).

Bugs reales corregidos:
- wp_insert_post/wp_update_post sin $wp_error=true: en runtime WP devuelve 0
  (no WP_Error) al fallar → el manejo is_wp_error era código muerto y seguía
  con update_post_meta(0,...) silenciosamente. Añadido true en 8 call sites.
- Email_Manager::build_variables() usaba $extra_vars sin definirlo (siempre
  '—'); ahora recibe el parámetro desde send().
- Cron_Manager: bloque method_exists(send_renovacion_fallida) eliminado — el
  método no existe y el email nunca se enviaba (fallo ya logueado).
- convoca-members.php: eliminados 3 'new CPT_*()' sin efecto (los CPT se
  registran vía add_action en init, líneas 277-279).
- webhooks: rama 'danger'/'error' muerta (el map solo tiene success/warning/info).
- PDF_Card::get_panel_url() eliminado (método privado sin uso).
- Casts esc_attr/esc_html int→string, paginación ceil float→int,
  meta_value int→string en get_posts, ternarios de wp_get_current_user
  simplificados (WP_User nunca null), ?? redundantes eliminados.

Verificado: PHPUnit 92 tests OK, PHPCS exit 0.

+++ admin/class-admin-logs.php
@@ -61,7 +61,7 @@ class Admin_Logs {
 				'%Members%'
 			)
 		);
-		$num_pages   = ceil( $total_items / $per_page );
+		$num_pages   = (int) ceil( $total_items / $per_page );
 
 		$logs = $wpdb->get_results(
 			$wpdb->prepare(
